settings refactor
